[v0.90][WP-14] Split instrumentation responsibilities (Closes #2051)
