+++ adl/src/instrumentation.rs
@@ -8,24 +8,15 @@ use serde::{Deserialize, Serialize};
 use crate::execution_plan::ExecutionPlan;
 use crate::trace::TraceEvent;
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct GraphNode {
-    pub id: String,
-    pub save_as: Option<String>,
-}
+mod graph;
+mod trace_formatting;
+mod trace_normalization;
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-pub struct GraphEdge {
-    pub from: String,
-    pub to: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct GraphExport {
-    pub workflow_kind: String,
-    pub nodes: Vec<GraphNode>,
-    pub edges: Vec<GraphEdge>,
-}
+pub use graph::{
+    export_graph, export_graph_dot, export_graph_json, GraphEdge, GraphExport, GraphNode,
+};
+pub use trace_formatting::format_normalized_event;
+pub use trace_normalization::normalize_trace_events;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 #[serde(tag = "kind", deny_unknown_fields)]
@@ -184,72 +175,6 @@ struct ActivationLogArtifact {
     ordering: String,
     stable_ids: ActivationLogStableIds,
     events: Vec<TraceEventNormalized>,
-}
-
-pub fn export_graph(plan: &ExecutionPlan) -> GraphExport {
-    let mut nodes: Vec<GraphNode> = plan
-        .nodes
-        .iter()
-        .map(|n| GraphNode {
-            id: n.step_id.clone(),
-            save_as: n.save_as.clone(),
-        })
-        .collect();
-    nodes.sort_by(|a, b| a.id.cmp(&b.id));
-
-    let mut edges: Vec<GraphEdge> = Vec::new();
-    for node in &plan.nodes {
-        for dep in &node.depends_on {
-            edges.push(GraphEdge {
-                from: dep.clone(),
-                to: node.step_id.clone(),
-            });
-        }
-    }
-    edges.sort();
-    edges.dedup();
-
-    GraphExport {
-        workflow_kind: format!("{:?}", plan.workflow_kind).to_lowercase(),
-        nodes,
-        edges,
-    }
-}
-
-pub fn export_graph_json(plan: &ExecutionPlan) -> Result<String> {
-    let graph = export_graph(plan);
-    serde_json::to_string_pretty(&graph).context("serialize graph json")
-}
-
-pub fn export_graph_dot(plan: &ExecutionPlan) -> String {
-    let graph = export_graph(plan);
-    let mut out = String::new();
-    out.push_str("digraph execution_plan {\n");
-    out.push_str("  rankdir=LR;\n");
-
-    for node in &graph.nodes {
-        let label = if let Some(save_as) = node.save_as.as_deref() {
-            format!("{}\\nsave_as={}", escape_dot(&node.id), escape_dot(save_as))
-        } else {
-            escape_dot(&node.id)
-        };
-        out.push_str(&format!(
-            "  \"{}\" [label=\"{}\"];\n",
-            escape_dot(&node.id),
-            label
-        ));
-    }
-
-    for edge in &graph.edges {
-        out.push_str(&format!(
-            "  \"{}\" -> \"{}\";\n",
-            escape_dot(&edge.from),
-            escape_dot(&edge.to)
-        ));
-    }
-
-    out.push_str("}\n");
-    out
 }
 
 pub fn write_trace_artifact(path: &Path, events: &[TraceEvent]) -> Result<()> {
@@ -450,321 +375,6 @@ pub fn diff_traces(left: &[TraceEventNormalized], right: &[TraceEventNormalized]
         left_only,
         right_only,
     }
-}
-
-pub fn normalize_trace_events(events: &[TraceEvent]) -> Vec<TraceEventNormalized> {
-    events
-        .iter()
-        .map(|ev| match ev {
-            TraceEvent::LifecyclePhaseEntered { phase, .. } => {
-                TraceEventNormalized::LifecyclePhaseEntered {
-                    phase: phase.as_str().to_string(),
-                }
-            }
-            TraceEvent::ExecutionBoundaryCrossed {
-                boundary, state, ..
-            } => TraceEventNormalized::ExecutionBoundaryCrossed {
-                boundary: boundary.as_str().to_string(),
-                state: state.clone(),
-            },
-            TraceEvent::SchedulerPolicy {
-                max_concurrency,
-                source,
-                ..
-            } => TraceEventNormalized::SchedulerPolicy {
-                max_concurrency: *max_concurrency,
-                source: source.clone(),
-            },
-            TraceEvent::RunFailed { message, .. } => TraceEventNormalized::RunFailed {
-                message: message.clone(),
-            },
-            TraceEvent::RunFinished { success, .. } => {
-                TraceEventNormalized::RunFinished { success: *success }
-            }
-            TraceEvent::StepStarted {
-                step_id,
-                agent_id,
-                provider_id,
-                task_id,
-                delegation,
-                ..
-            } => TraceEventNormalized::StepStarted {
-                step_id: step_id.clone(),
-                agent_id: agent_id.clone(),
-                provider_id: provider_id.clone(),
-                task_id: task_id.clone(),
-                delegation_json: delegation.as_ref().and_then(|d| {
-                    if d.is_effectively_empty() {
-                        None
-                    } else {
-                        serde_json::to_string(&d.canonicalized()).ok()
-                    }
-                }),
-            },
-            TraceEvent::PromptAssembled {
-                step_id,
-                prompt_hash,
-                ..
-            } => TraceEventNormalized::PromptAssembled {
-                step_id: step_id.clone(),
-                prompt_hash: prompt_hash.clone(),
-            },
-            TraceEvent::StepOutputChunk {
-                step_id,
-                chunk_bytes,
-                ..
-            } => TraceEventNormalized::StepOutputChunk {
-                step_id: step_id.clone(),
-                chunk_bytes: *chunk_bytes,
-            },
-            TraceEvent::DelegationRequested {
-                delegation_id,
-                step_id,
-                action_kind,
-                target_id,
-                ..
-            } => TraceEventNormalized::DelegationRequested {
-                delegation_id: delegation_id.clone(),
-                step_id: step_id.clone(),
-                action_kind: action_kind.clone(),
-                target_id: target_id.clone(),
-            },
-            TraceEvent::DelegationPolicyEvaluated {
-                delegation_id,
-                step_id,
-                action_kind,
-                target_id,
-                decision,
-                rule_id,
-                ..
-            } => TraceEventNormalized::DelegationPolicyEvaluated {
-                delegation_id: delegation_id.clone(),
-                step_id: step_id.clone(),
-                action_kind: action_kind.clone(),
-                target_id: target_id.clone(),
-                decision: decision.clone(),
-                rule_id: rule_id.clone(),
-            },
-            TraceEvent::DelegationApproved {
-                delegation_id,
-                step_id,
-                ..
-            } => TraceEventNormalized::DelegationApproved {
-                delegation_id: delegation_id.clone(),
-                step_id: step_id.clone(),
-            },
-            TraceEvent::DelegationDenied {
-                delegation_id,
-                step_id,
-                action_kind,
-                target_id,
-                rule_id,
-                ..
-            } => TraceEventNormalized::DelegationDenied {
-                delegation_id: delegation_id.clone(),
-                step_id: step_id.clone(),
-                action_kind: action_kind.clone(),
-                target_id: target_id.clone(),
-                rule_id: rule_id.clone(),
-            },
-            TraceEvent::DelegationDispatched {
-                delegation_id,
-                step_id,
-                action_kind,
-                target_id,
-                ..
-            } => TraceEventNormalized::DelegationDispatched {
-                delegation_id: delegation_id.clone(),
-                step_id: step_id.clone(),
-                action_kind: action_kind.clone(),
-                target_id: target_id.clone(),
-            },
-            TraceEvent::DelegationResultReceived {
-                delegation_id,
-                step_id,
-                success,
-                output_bytes,
-                ..
-            } => TraceEventNormalized::DelegationResultReceived {
-                delegation_id: delegation_id.clone(),
-                step_id: step_id.clone(),
-                success: *success,
-                output_bytes: *output_bytes,
-            },
-            TraceEvent::DelegationCompleted {
-                delegation_id,
-                step_id,
-                outcome,
-                ..
-            } => TraceEventNormalized::DelegationCompleted {
-                delegation_id: delegation_id.clone(),
-                step_id: step_id.clone(),
-                outcome: outcome.clone(),
-            },
-            TraceEvent::StepFinished {
-                step_id, success, ..
-            } => TraceEventNormalized::StepFinished {
-                step_id: step_id.clone(),
-                success: *success,
-            },
-            TraceEvent::CallEntered {
-                caller_step_id,
-                callee_workflow_id,
-                namespace,
-                ..
-            } => TraceEventNormalized::CallEntered {
-                caller_step_id: caller_step_id.clone(),
-                callee_workflow_id: callee_workflow_id.clone(),
-                namespace: namespace.clone(),
-            },
-            TraceEvent::CallExited {
-                caller_step_id,
-                status,
-                namespace,
-                ..
-            } => TraceEventNormalized::CallExited {
-                caller_step_id: caller_step_id.clone(),
-                status: status.clone(),
-                namespace: namespace.clone(),
-            },
-        })
-        .collect()
-}
-
-pub fn format_normalized_event(ev: &TraceEventNormalized) -> String {
-    match ev {
-        TraceEventNormalized::LifecyclePhaseEntered { phase } => {
-            format!("LifecyclePhaseEntered phase={phase}")
-        }
-        TraceEventNormalized::ExecutionBoundaryCrossed { boundary, state } => {
-            format!("ExecutionBoundaryCrossed boundary={boundary} state={state}")
-        }
-        TraceEventNormalized::SchedulerPolicy {
-            max_concurrency,
-            source,
-        } => format!("SchedulerPolicy max_concurrency={max_concurrency} source={source}"),
-        TraceEventNormalized::RunFailed { message } => {
-            format!("RunFailed message={message}")
-        }
-        TraceEventNormalized::RunFinished { success } => {
-            format!("RunFinished success={success}")
-        }
-        TraceEventNormalized::StepStarted {
-            step_id,
-            agent_id,
-            provider_id,
-            task_id,
-            delegation_json,
-        } => {
-            let base = format!(
-                "StepStarted step={step_id} agent={agent_id} provider={provider_id} task={task_id}"
-            );
-            if let Some(d) = delegation_json {
-                format!("{base} delegation={d}")
-            } else {
-                base
-            }
-        }
-        TraceEventNormalized::PromptAssembled {
-            step_id,
-            prompt_hash,
-        } => {
-            format!("PromptAssembled step={step_id} hash={prompt_hash}")
-        }
-        TraceEventNormalized::StepOutputChunk {
-            step_id,
-            chunk_bytes,
-        } => {
-            format!("StepOutputChunk step={step_id} bytes={chunk_bytes}")
-        }
-        TraceEventNormalized::DelegationRequested {
-            delegation_id,
-            step_id,
-            action_kind,
-            target_id,
-        } => format!(
-            "DelegationRequested delegation_id={delegation_id} step={step_id} action={action_kind} target={target_id}"
-        ),
-        TraceEventNormalized::DelegationPolicyEvaluated {
-            delegation_id,
-            step_id,
-            action_kind,
-            target_id,
-            decision,
-            rule_id,
-        } => {
-            let base = format!(
-                "DelegationPolicyEvaluated delegation_id={delegation_id} step={step_id} action={action_kind} target={target_id} decision={decision}"
-            );
-            if let Some(rule_id) = rule_id {
-                format!("{base} rule_id={rule_id}")
-            } else {
-                base
-            }
-        }
-        TraceEventNormalized::DelegationApproved {
-            delegation_id,
-            step_id,
-        } => format!("DelegationApproved delegation_id={delegation_id} step={step_id}"),
-        TraceEventNormalized::DelegationDenied {
-            delegation_id,
-            step_id,
-            action_kind,
-            target_id,
-            rule_id,
-        } => {
-            let base = format!(
-                "DelegationDenied delegation_id={delegation_id} step={step_id} action={action_kind} target={target_id}"
-            );
-            if let Some(rule_id) = rule_id {
-                format!("{base} rule_id={rule_id}")
-            } else {
-                base
-            }
-        }
-        TraceEventNormalized::DelegationDispatched {
-            delegation_id,
-            step_id,
-            action_kind,
-            target_id,
-        } => format!(
-            "DelegationDispatched delegation_id={delegation_id} step={step_id} action={action_kind} target={target_id}"
-        ),
-        TraceEventNormalized::DelegationResultReceived {
-            delegation_id,
-            step_id,
-            success,
-            output_bytes,
-        } => format!(
-            "DelegationResultReceived delegation_id={delegation_id} step={step_id} success={success} bytes={output_bytes}"
-        ),
-        TraceEventNormalized::DelegationCompleted {
-            delegation_id,
-            step_id,
-            outcome,
-        } => format!("DelegationCompleted delegation_id={delegation_id} step={step_id} outcome={outcome}"),
-        TraceEventNormalized::StepFinished { step_id, success } => {
-            format!("StepFinished step={step_id} success={success}")
-        }
-        TraceEventNormalized::CallEntered {
-            caller_step_id,
-            callee_workflow_id,
-            namespace,
-        } => {
-            format!("CallEntered caller_step={caller_step_id} callee_workflow={callee_workflow_id} namespace={namespace}")
-        }
-        TraceEventNormalized::CallExited {
-            caller_step_id,
-            status,
-            namespace,
-        } => {
-            format!("CallExited caller_step={caller_step_id} status={status} namespace={namespace}")
-        }
-    }
-}
-
-fn escape_dot(v: &str) -> String {
-    v.replace('\\', "\\\\").replace('"', "\\\"")
 }
 
 #[cfg(test)]
